PS1: Refactored todo controller
